add additional mm tests
